# Use fewer Phlex::Blocks