--- a/spec/phlex/component_spec.rb
+++ b/spec/phlex/component_spec.rb
@@ -377,20 +377,6 @@ RSpec.describe Phlex::Component do
     end
   end
 
-  describe "with a positional argument" do
-    let(:component) do
-      Class.new Phlex::Component do
-        def template
-          component CardComponent, "Hello World!"
-        end
-      end
-    end
-
-    it "produces the correct markup" do
-      expect(output).to eq %{<article class="p-5 rounded drop-shadow">Hello World!</article>}
-    end
-  end
-
   describe "#format" do
     it "returns :html" do
       expect(example.format).to eq :html