I realised we don't need to use so many `Phlex::Blocks`. An 86% performance improvement.

Before:
```
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [arm64-darwin21]
Warming up --------------------------------------
                Page    40.000  i/100ms
Calculating -------------------------------------
                Page    398.825  (± 0.5%) i/s -      2.000k in   5.014892s
```

```
               phlex:     8095.3 i/s
      view_component:     5362.8 i/s - 1.51x  (± 0.00) slower
               cells:     3458.0 i/s - 2.34x  (± 0.00) slower
            partials:     2481.5 i/s - 3.26x  (± 0.00) slower
            dry_view:      342.2 i/s - 23.66x  (± 0.00) slower
```


After:
```
ruby 3.1.2p20 (2022-04-12 revision 4491bb740a) [arm64-darwin21]
Warming up --------------------------------------
                Page    74.000  i/100ms
Calculating -------------------------------------
                Page    742.816  (± 0.1%) i/s -      3.774k in   5.080680s
```

```
               phlex:    10571.1 i/s
      view_component:     5395.3 i/s - 1.96x  (± 0.00) slower
               cells:     3491.6 i/s - 3.03x  (± 0.00) slower
            partials:     2501.6 i/s - 4.23x  (± 0.00) slower
            dry_view:      344.9 i/s - 30.65x  (± 0.00) slower
```